% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arousal_Simulation_No_Stochasticity.R
\name{Arousal_Simulation_No_Stochasticity}
\alias{Arousal_Simulation_No_Stochasticity}
\title{Generative model of Arousal based turn-taking.}
\usage{
Arousal_Simulation_No_Stochasticity(n, mu_latency, sd_latency)
}
\arguments{
\item{n}{Number of Observations}

\item{mu_latency}{Average latency between consecutive calls per caller.}

\item{sd_latency}{Uncertainty around the average latency between calls.}
}
\value{
A data frame object that contains our dyadic interlocutors
}
\description{
Simulates dyadic vocal turn taking timings based on the arousal model.
Calls have no duration.
}
\examples{

}
