% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arousal_Simulation_Stochastic.R
\name{Arousal_Simulation_Stochastic}
\alias{Arousal_Simulation_Stochastic}
\title{Generative model of Arousal based turn-taking.}
\usage{
Arousal_Simulation_Stochastic(n, mu_latency, sd_latency)
}
\arguments{
\item{n}{Number of Observations}

\item{mu_latency}{Average latency between consecutive calls per caller.}

\item{sd_latency}{Uncertainty around the average latency between calls.}
}
\value{
A data frame object that contains our dyadic interlocutors
}
\description{
Simulates dyadic vocal turn taking timings based on the arousal model.
This model includes a stochastic element. The probability of Interlocutor B  calling increases as the number of calls heard from
interlocutor B increases. We have modelled the relationship with a sigmoid activation function.
Calls have no duration.
}
\examples{

}
