% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InhibitionHypothesis.R
\name{InhibitionHypothesis}
\alias{InhibitionHypothesis}
\title{Inhibition Hypothesis Testing}
\usage{
InhibitionHypothesis(df)
}
\arguments{
\item{df}{The data frame we wish to test.}
}
\value{
Summary of the test output.
}
\description{
We test whether empirical data could have been generated under the inhibition hypothesis.
}
\examples{

}
