% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reset_Simulation.R
\name{Reset_Simulation}
\alias{Reset_Simulation}
\title{Baseline Simulation}
\usage{
Reset_Simulation(n, mu_latency, sd_latency, mu_duration, sd_duration)
}
\arguments{
\item{n}{Number of Observations}

\item{mu_latency}{Average latency between consecutive calls per caller.}

\item{sd_latency}{Uncertainty around the average latency between calls.}

\item{mu_duration}{Average duration of each call.}

\item{sd_duration}{Uncertainty in average call duration.}
}
\value{
A data frame object that contains our dyadic interlocutors
}
\description{
Simulates baseline dyadic vocal turn taking based on the reset hypothesis.
All callers have an internal clock which determines the latency between consecutive calls.
The clock of caller B is reset if caller A is calling during the period in which caller B were expected to be calling.
}
\examples{

}
