% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetHypothesis.R
\name{ResetHypothesis}
\alias{ResetHypothesis}
\title{Reset Hypothesis Testing}
\usage{
ResetHypothesis(data)
}
\arguments{
\item{df}{The data frame we wish to test.}
}
\value{
Summary of the test output.
}
\description{
We test whether empirical data could have been generated under the reset hypothesis.
}
\examples{

}
