% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotateData.R
\name{RotateData}
\alias{RotateData}
\title{Rotate Data}
\usage{
RotateData(d, threshold, delay)
}
\arguments{
\item{d}{The data frame we wish to rotate.}

\item{threshold}{Threshold for cutoff}

\item{delay}{Delay duration of the snippet being cut.}
}
\value{
A data frame containing the rotated time series.
}
\description{
We create rotated data sets for each caller. First we add a random n of seconds to the time series at the start.
   We use the shifted time series which are now longer than the bout and we cut and paste them together.
}
\examples{

}
